pyproject.toml
[project]
name = "dbt-mcp"
version = "0.1.0"
description = "A MCP (Model Context Protocol) server for interacting with dbt resources."
authors = [{ name = "dbt Labs" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
classifiers = [
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
]
dependencies = [
  "dbt-sl-sdk[sync]>=0.10.0",
  # We are using some internal functions, so pinning this dependency.
  "mcp[cli]==1.6.0",
  "pandas>=2.2.3",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
]
[dependency-groups]
dev = [
  "ruff>=0.11.2",
  "types-requests>=2.32.0.20250328",
  "mypy>=1.12.1",
  "pre-commit>=4.2.0",
  "pytest-asyncio>=0.26.0",
  "pytest>=8.3.5",
  "openai>=1.71.0",
  "pyarrow-stubs>=19.1",
]

[project.urls]
Documentation = "https://github.com/dbt-labs/dbt-mcp#readme"
Issues = "https://github.com/dbt-labs/dbt-mcp/issues"
Source = "https://github.com/dbt-labs/dbt-mcp"

[project.scripts]
dbt-mcp = "dbt_mcp.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dbt_mcp"]

[tool.hatch.build.targets.sdist]
include = ["./dbt_mcp", "./README.md", "./LICENSE"]
